% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggimage.r
\name{ggimage}
\alias{ggimage}
\title{A ggplot2 based raster image function}
\usage{
ggimage(
  zmat,
  x = NULL,
  y = NULL,
  zlim = range(zmat, na.rm = TRUE),
  col = viridisLite::viridis(256),
  xlab = NULL,
  ylab = NULL,
  show.legend = TRUE,
  legend = NULL,
  title = NULL,
  xrev = FALSE,
  yrev = FALSE,
  asp = 1,
  addcontour = FALSE,
  binwidth = NULL
)
}
\arguments{
\item{zmat}{either the matrix to display or a names list with elements ra, dec, and zmat}

\item{x}{x axis values -- either a vector the same length as number of rows or a matrix with the same dimensions as `zmat`}

\item{y}{y axis values or NULL}

\item{zlim}{range of z values to display}

\item{col}{color palette. Defaults to viridis(256)}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{show.legend}{display legend?}

\item{legend}{label for legend}

\item{title}{main graph title}

\item{xrev}{if `TRUE` reverse X axis}

\item{yrev}{if `TRUE` reverse Y axis}

\item{asp}{aspect ratio}

\item{addcontour}{add contours?}

\item{binwidth}{contour spacing if used}
}
\value{
data frame constructed to build plot and the graph, invisibly in a list
}
\description{
Displays a matrix as a raster image with user selectable colors
}
\details{
`zmat` is the maatrix to display. All other arguments are optional
with mostly `NULL` defaults
}
