% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greyimage.r
\docType{data}
\name{grey256}
\alias{grey256}
\alias{gray256}
\title{grey256}
\format{
An object of class \code{character} of length 256.
}
\usage{
grey256
}
\description{
256 level greyscale palette
}
\keyword{datasets}
