% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greyimage.r
\name{greyimage}
\alias{greyimage}
\alias{grayimage}
\title{base graphics raster image}
\usage{
greyimage(
  zmat,
  x = NULL,
  y = NULL,
  col = grey(seq(0, 1, length = 256)),
  xlab = "X",
  ylab = "Y",
  ...
)
}
\arguments{
\item{zmat}{matrix to be displayed, or a named list with the image, x coordinates, y coordinates}

\item{x}{vector of x coordinates}

\item{y}{vector of y coordinates}

\item{col}{color palette}

\item{xlab}{X axis label}

\item{ylab}{Y axis label}

\item{...}{additional arguments passed to `image()`}
}
\description{
A simple raster image function with sensible defaults
}
\details{
The color palette defaults to 256 level grayscale, but can be anything
}
