\name{myggplot-package}
\alias{myggplot-package}
\alias{myggplot}
\docType{package}
\title{
\packageTitle{myggplot}
}
\description{
\packageDescription{myggplot}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{myggplot}
\packageIndices{myggplot}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{myggplot}

Maintainer: \packageMaintainer{myggplot}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
