% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot.r
\name{qplot}
\alias{qplot}
\title{my pilfered `qplot`}
\usage{
qplot(
  x,
  y,
  ...,
  data,
  facets = NULL,
  margins = FALSE,
  geom = "auto",
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  log = "",
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  asp = NA
)
}
\arguments{
\item{x}{x values to plot}

\item{y}{y values}

\item{...}{other variables in plot}

\item{facets}{facets}

\item{margins}{margins}

\item{geom}{`geom` to use or `auto` to make a reasonable guess}

\item{xlim}{x axis limits}

\item{ylim}{y axis limits}

\item{log}{specify axis to be log-scaled}

\item{main}{graph title}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{asp}{x/y aspect ratio}
}
\value{
the graph
}
\description{
ggplot2's qplot function without the annoying deprecated warnings
}
\details{
This was copied directly from the ggplot2 function of the same name. Checks for
deprecation were edited out. Also, some unexported ggplot2 functions are invoked
with `ggplot2:::`.
}
\examples{
x <- runif(100)
y <- x + rnorm(100, sd=0.1)
qplot(x, y, main="Fake data")

}
