% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greyimage.r
\name{greyimage}
\alias{greyimage}
\alias{grayimage}
\title{base graphics raster image}
\usage{
greyimage(img, col = grey256, xlab = "X", ylab = "Y", ...)
}
\arguments{
\item{img}{matrix to be displayed}

\item{col}{color palette}

\item{xlab}{X axis label}

\item{ylab}{Y axis label}

\item{...}{additional arguments passed to `image()`}
}
\description{
A simple raster image function with sensible defaults
}
\details{
The color palette defaults to 256 level grayscale, but can be anything
}
